Extract opt in redirect logic into separate controller, and fix Okta login on Firefox / Safari
